Removed reducer and used SortableStatsContext
