+++ src/modules/pools/pod-staking/index.jsx
@@ -13,11 +13,11 @@ import { CardSkeleton } from "@/common/Skeleton/CardSkeleton";
 import { CARDS_PER_PAGE } from "@/src/config/constants";
 import { PodStakingCard } from "@/src/modules/pools/pod-staking/PodStakingCard";
 import { usePodStakingPools } from "@/src/hooks/usePodStakingPools";
-import { useStakingPoolsStats } from "@/modules/pools/staking/StakingPoolsStatsContext";
+import { useSortableStats } from "@/src/context/SortableStatsContext";
 import { toStringSafe } from "@/utils/string";
 
 /**
- * @type {Object.<string, {selector:(any) => any, datatype: any }>}
+ * @type {Object.<string, {selector:(any) => any, datatype: any, ascending?: boolean }>}
  */
 const sorterData = {
   [SORT_TYPES.ALPHABETIC]: {
@@ -42,7 +42,7 @@ export const PodStakingPage = () => {
   const router = useRouter();
 
   const { data, loading, hasMore, handleShowMore } = usePodStakingPools();
-  const { getStatsByKey } = useStakingPoolsStats();
+  const { getStatsByKey } = useSortableStats();
   const { getTVLById } = useAppConstants();
 
   const { searchValue, setSearchValue, filtered } = useSearchResults({
