+++ src/modules/reporting/resolved/ResolvedReportingCard.jsx
@@ -1,3 +1,6 @@
+import { useEffect } from "react";
+import { useRouter } from "next/router";
+
 import { Divider } from "@/common/Divider/Divider";
 import { OutlinedCard } from "@/common/OutlinedCard/OutlinedCard";
 import { getCoverImgSrc } from "@/src/helpers/cover";
@@ -6,12 +9,20 @@ import { fromNow } from "@/utils/formatter/relative-time";
 import DateLib from "@/lib/date/DateLib";
 import { CardStatusBadge } from "@/common/CardStatusBadge";
 import { Trans } from "@lingui/macro";
-import { useRouter } from "next/router";
+import { useSortableStats } from "@/src/context/SortableStatsContext";
 
-export const ResolvedReportingCard = ({ coverKey, status, resolvedOn }) => {
+export const ResolvedReportingCard = ({ id, coverKey, status, resolvedOn }) => {
+  const { setStatsByKey } = useSortableStats();
   const { coverInfo } = useCoverInfo(coverKey);
   const imgSrc = getCoverImgSrc({ key: coverKey });
   const router = useRouter();
+
+  // Used for sorting purpose only
+  useEffect(() => {
+    setStatsByKey(id, {
+      resolvedOn,
+    });
+  }, [id, resolvedOn, setStatsByKey]);
 
   return (
     <OutlinedCard className="p-6 bg-white" type="link">
@@ -36,7 +47,10 @@ export const ResolvedReportingCard = ({ coverKey, status, resolvedOn }) => {
 
       {/* Stats */}
       <div className="flex justify-between px-1 mb-4 text-sm">
-        <span className="" title={DateLib.toLongDateFormat(resolvedOn, router.locale)}>
+        <span
+          className=""
+          title={DateLib.toLongDateFormat(resolvedOn, router.locale)}
+        >
           <Trans>Resolved On:</Trans>{" "}
           <span title={DateLib.toLongDateFormat(resolvedOn, router.locale)}>
             {fromNow(resolvedOn)}
