+++ src/pages/pools/pod-staking.jsx
@@ -4,7 +4,7 @@ import { PoolsTabs } from "@/src/modules/pools/PoolsTabs";
 import { PodStakingPage } from "@/src/modules/pools/pod-staking";
 import { ComingSoon } from "@/common/ComingSoon";
 import { isFeatureEnabled } from "@/src/config/environment";
-import { StakingStatsProvider } from "@/modules/pools/staking/StakingPoolsStatsContext";
+import { SortableStatsProvider } from "@/src/context/SortableStatsContext";
 
 export function getStaticProps() {
   return {
@@ -29,9 +29,9 @@ export default function PodStaking({ disabled }) {
         />
       </Head>
       <PoolsTabs active="pod-staking">
-        <StakingStatsProvider>
+        <SortableStatsProvider>
           <PodStakingPage />
-        </StakingStatsProvider>
+        </SortableStatsProvider>
       </PoolsTabs>
     </main>
   );
