+++ src/pages/reporting/active.jsx
@@ -3,6 +3,7 @@ import { ReportingTabs } from "@/src/modules/reporting/ReportingTabs";
 import { ReportingActivePage } from "@/src/modules/reporting/active/active";
 import { ComingSoon } from "@/common/ComingSoon";
 import { isFeatureEnabled } from "@/src/config/environment";
+import { SortableStatsProvider } from "@/src/context/SortableStatsContext";
 
 export function getStaticProps() {
   return {
@@ -27,7 +28,9 @@ export default function ReportingActive({ disabled }) {
         />
       </Head>
       <ReportingTabs active="active">
-        <ReportingActivePage />
+        <SortableStatsProvider>
+          <ReportingActivePage />
+        </SortableStatsProvider>
       </ReportingTabs>
     </main>
   );
